requirements.txt
amqp==5.3.1
asgiref==3.11.0
async-timeout==5.0.1
billiard==4.2.4
CacheControl==0.14.4
celery==5.3.4
certifi==2026.1.4
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron_descriptor==2.0.6
cryptography==46.0.3
Django==4.2.7
django-celery-beat==2.5.0
django-environ==0.11.2
django-timezone-field==7.2.1
djangorestframework==3.14.0
firebase-admin
google-api-core
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-generativeai
google-resumable-media
googleapis-common-protos>=1.56.0
grpcio>=1.59.0,<1.63.0
grpcio-status>=1.59.0,<1.63.0
httplib2==0.31.0
idna==3.11
kombu==5.6.2
msgpack==1.1.2
packaging==25.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf>=4.25.3,<5.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.10.1
pyparsing==3.3.1
python-crontab==3.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
redis==5.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.14
